+++ test.mjs
@@ -981,6 +981,69 @@ jstestDescribe((
         });
     });
     jstestIt((
+        "test sqlite-extension-jfromfloat64array handling-behavior"
+    ), async function test_sqliteExtensionJfromfloat64array() {
+        let db = await dbOpenAsync({
+            filename: ":memory:"
+        });
+        [
+            [" [ , 1 ] ", "error"],
+            [" [ , ] ", "error"],
+            [" [ 1 , ] ", "error"],
+            [" [ ] ", "[]"],
+            [" [ null ] ", "[0.0]"],
+            ["", "error"],
+            ["1,2]", "error"],
+            ["[,1]", "error"],
+            ["[,]", "error"],
+            ["[0]", "[0.0]"],
+            ["[1,2", "error"],
+            ["[1,2,a]", "[1.0,2.0,0.0]"],
+            ["[1,]", "error"],
+            ["[1,a,3]", "[1.0,0.0,3.0]"],
+            ["[1]", "[1.0]"],
+            ["[]", "[]"],
+            ["[a,2,3]", "[0.0,2.0,3.0]"],
+            [0, "error"],
+            [1, "error"],
+            [`[${"1".repeat(100)}]`, `[1.11111111111111e+99]`],
+            [null, "error"],
+            [undefined, "error"],
+            [{}, "error"]
+        ].forEach(async function ([
+            valInput, valExpected
+        ], ii) {
+            let valActual;
+            try {
+                valActual = noop(
+                    await dbExecAsync({
+                        bindList: {
+                            valInput
+                        },
+                        db,
+                        sql: (`
+SELECT jfromfloat64array(jtofloat64array($valInput)) AS result;
+                        `)
+                    })
+                )[0][0].result;
+            } catch (ignore) {
+                assertOrThrow(valExpected === "error", JSON.stringify({
+                    ii,
+                    valActual,
+                    valExpected,
+                    valInput
+                }, undefined, 4));
+                return;
+            }
+            assertJsonEqual(valActual, valExpected, {
+                ii,
+                valActual,
+                valExpected,
+                valInput
+            });
+        });
+    });
+    jstestIt((
         "test sqlite-extension-kthpercentile handling-behavior"
     ), async function test_sqliteExtensionKthpercentile() {
         let db = await dbOpenAsync({